start now button and some refactoring
