+++ src/pages/StartNowPage.vue
